+++ src/Entity/IterableFieldTrait.php
@@ -4,59 +4,14 @@ declare(strict_types=1);
 
 namespace Bolt\Entity;
 
-trait IterableFieldTrait
-{
-    private $iteratorCursor = 0;
-
-    /** @var array Field */
-    private $fields = [];
-
-    /**
-     * Makes ListFieldInterface fields |length filter
-     * return the number of elements in the field
-     */
-    public function count(): int
+if (PHP_MAJOR_VERSION >= 8) {
+    trait IterableFieldTrait
     {
-        return count($this->getValue());
+        use IterableFieldTraitPhp8;
     }
-
-    /**
-     * Makes ListFieldInterface fields .length attribute
-     * return the number of elements in the field
-     */
-    public function length(): int
+} else {
+    trait IterableFieldTrait
     {
-        return $this->count();
-    }
-
-    /**
-     * @return Field|string
-     */
-    public function current()
-    {
-        return $this->fields[$this->iteratorCursor];
-    }
-
-    public function next(): void
-    {
-        ++$this->iteratorCursor;
-    }
-
-    public function key(): int
-    {
-        return $this->iteratorCursor;
-    }
-
-    public function valid(): bool
-    {
-        return isset($this->fields[$this->iteratorCursor]);
-    }
-
-    public function rewind(): void
-    {
-        // Ensure $this->fields is initialised
-        $this->fields = $this->getValue();
-
-        $this->iteratorCursor = 0;
+        use IterableFieldTraitPhp7;
     }
 }
